Disable ncurses for oksh
When moving all builds to a single Docker stage, the shells started
to build all in the same machine, causing oksh to be linked against
ncurses.

Since the build stage is different from the final image, ncurses
was missing and causing oksh binaries to fail invokation.

This change disables linking against ncurses during build, fixing
the issue.

Our goal is not to provide interactive shells for personal use,
so lacking ncurses still results in a perfectly good oksh.