Support bash 2.05b
The old bash 2.05b is not used in typical production machines, but
it is common amongst bootstrapping tools and it is considered the
oldest shell with modern features.

Although we currently prefer using package-provided configure
scripts, even when autoconf is available, this was not possible
for bash 2.05b, which required a regenerated configure script.