Add rust shells: yash-rs and brush
These shells have achieved some significant compatibility with
more estabilished interpreters and are being included in the
build.

Although rust yash is made by the same author and the binary
is named yash3, we decided to keep it separately from the
already existing yash shell due to version numbering
differences and being a completely new codebase.