Upgrade Debian from buster to bullseye
In this project we use older base systems, in order to keep the
patches to upstream packages as few as possible.

The buster version of Debian was the oldest available that was
still maintained, but it has reached end-of-life.

In this commit, we upgrade to the next thing, debian bullseye.