New bash, dash, osh, posh, yash, yashrs versions. Add brush.
yashrs is now using a versioning scheme that implies it should
be considered a compatible rewrite of yash, but in rust, moving
from 2.x to 3x. However, earlier versions with the older versioning
scheme are still in our horizon for providing support. As the 3.x
series get more releases, we should merge it back to a single `yash`
identifier.

Notably, dash and posh got releases after a lot of years without an
update.

We are also introducing back historical beta and rc releases for
bash, since people might be interested in testing them.

This closes issue #21 regarding brush.