Disable `all` image, separate downloads and builds, cache downloads
This change introduces separate downloads from the builds, in order
to cache them on the CI process.

It also introduces automatic regeneration of GitHub workflow YAML
files, so we don't have to maintain them for changing versions
anymore.

Due to constraints on the CI, we are disabling the generation of
the `all` image, leaving only `latest` to be built. The full
image can still be built locally by whoever wants it.